requirements.txt
django
qrcode
pillow
python-dotenv
